+++ tests/Unit/ResolveSocialiteUsersTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace JoelButcher\Socialstream\Tests;
+namespace JoelButcher\Socialstream\Tests\Unit;
 
 use JoelButcher\Socialstream\Contracts\ResolvesSocialiteUsers;
 use JoelButcher\Socialstream\Socialstream;
 use JoelButcher\Socialstream\Tests\Fixtures\ResolveUser;
+use JoelButcher\Socialstream\Tests\TestCase;
 
 class ResolveSocialiteUsersTest extends TestCase
 {
