+++ tests/Unit/SetPasswordTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace JoelButcher\Socialstream\Tests;
+namespace JoelButcher\Socialstream\Tests\Unit;
 
 use App\Actions\Socialstream\SetUserPassword;
 use Illuminate\Support\Facades\Hash;
 use JoelButcher\Socialstream\Tests\Fixtures\User;
+use JoelButcher\Socialstream\Tests\TestCase;
 
 class SetPasswordTest extends TestCase
 {
